refactor(eip5792): move batch predicate to shared tx-batch.hooks and wire EIP-7702
Move useShouldBatchTransactions and useStepWithTxBatch out of safe.hooks.tsx into a new shared tx-batch.hooks.ts module. The predicate now returns true for both Safe smart accounts and EIP-7702 wallets with atomic capability (supported/ready), gated by the same shouldUseTxBundling user setting. Update all importers across both apps to point to the new module. Clean up unused imports from safe.hooks.tsx. Update useSafeBatchSubmitter to import BatchSubmitter from the shared lib.
