+++ packages/lib/modules/pool/actions/add-liquidity/modal/AddLiquidityModal.tsx
@@ -18,7 +18,7 @@ import { useAddLiquidityReceipt } from '@repo/lib/modules/transactions/transacti
 import { useUserAccount } from '@repo/lib/modules/web3/UserAccountProvider'
 import { useTokens } from '@repo/lib/modules/tokens/TokensProvider'
 import { TxBatchAlert } from '@repo/lib/shared/components/alerts/TxBatchAlert'
-import { useShouldBatchTransactions } from '@repo/lib/modules/web3/safe.hooks'
+import { useShouldBatchTransactions } from '@repo/lib/modules/transactions/transaction-steps/tx-batch.hooks'
 import { ProtocolVersion } from '../../../pool.types'
 
 type Props = {
