+++ packages/lib/modules/pool/actions/create/modal/PoolCreationModal.tsx
@@ -9,7 +9,7 @@ import { Address } from 'viem'
 import { useRedirect } from '@repo/lib/shared/hooks/useRedirect'
 import { ActionModalFooter } from '@repo/lib/shared/components/modals/ActionModalFooter'
 import { getChainName } from '@repo/lib/config/app.config'
-import { useShouldBatchTransactions } from '@repo/lib/modules/web3/safe.hooks'
+import { useShouldBatchTransactions } from '@repo/lib/modules/transactions/transaction-steps/tx-batch.hooks'
 import { PoolSummary } from './PoolSummary'
 import { ToggleHyperBlockSize } from './ToggleHyperBlockSize'
 import { useHyperEvm } from '@repo/lib/modules/chains/hyperevm/useHyperEvm'
