+++ packages/lib/modules/pool/actions/create/modal/usePoolCreationTransactions.ts
@@ -9,7 +9,7 @@ import { useTokenApprovalSteps } from '@repo/lib/modules/tokens/approvals/useTok
 import { useIsPoolInitialized } from '@repo/lib/modules/pool/queries/useIsPoolInitialized'
 import { useSignPermit2InitializeStep } from '@repo/lib/modules/pool/actions/initialize/useSignPermit2InitializeStep'
 import { usePermit2ApprovalSteps } from '@repo/lib/modules/tokens/approvals/permit2/usePermit2ApprovalSteps'
-import { useShouldBatchTransactions } from '@repo/lib/modules/web3/safe.hooks'
+import { useShouldBatchTransactions } from '@repo/lib/modules/transactions/transaction-steps/tx-batch.hooks'
 import { useUserSettings } from '@repo/lib/modules/user/settings/UserSettingsProvider'
 import { getApprovalAndAddSteps } from '@repo/lib/modules/pool/actions/add-liquidity/useAddLiquiditySteps'
 import { useInitializePoolStep } from './useInitializePoolStep'
