+++ packages/lib/modules/pool/actions/remove-liquidity/useRemoveLiquiditySteps.tsx
@@ -5,7 +5,8 @@ import { TransactionStep } from '@repo/lib/modules/transactions/transaction-step
 import { useSignPermitStep } from '@repo/lib/modules/transactions/transaction-steps/useSignPermitStep'
 import { useSignRelayerStep } from '@repo/lib/modules/transactions/transaction-steps/useSignRelayerStep'
 import { useUserSettings } from '@repo/lib/modules/user/settings/UserSettingsProvider'
-import { useIsSafeAccount, useShouldBatchTransactions } from '@repo/lib/modules/web3/safe.hooks'
+import { useIsSafeAccount } from '@repo/lib/modules/web3/safe.hooks'
+import { useShouldBatchTransactions } from '@repo/lib/modules/transactions/transaction-steps/tx-batch.hooks'
 import { useMemo } from 'react'
 import { usePool } from '../../PoolProvider'
 import { isV3Pool } from '../../pool.helpers'
