+++ packages/lib/shared/components/modals/ActionModalFooter.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button, Divider, HStack, ModalFooter, VStack, Link } from '@chakra-ui/react'
-import { useStepWithTxBatch } from '@repo/lib/modules/web3/safe.hooks'
+import { useStepWithTxBatch } from '@repo/lib/modules/transactions/transaction-steps/tx-batch.hooks'
 import { AnimatePresence, motion } from 'motion/react'
 import { PropsWithChildren } from 'react'
 import { CornerDownLeft, MessageSquare } from 'lucide-react'
